Upstart-LatestNews-Section Done
The writings at the top of the section were made like the tops of other sections. The cards in the section were created and texts and pictures were added to it. Changes have also been made when hovering over the a tags in card structures and card structures. Added shadow to the card structure.